Refactor: Member UserDetails를 별도의 클래스에 분리 (#297)
